src/third_party/blink/web_tests/external/wpt/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/url.window-expected.txt

This is a testharness.js-based test.
PASS document.open() changes document's URL (fully active document)
FAIL document.open() does not change document's URL (active but not fully active document) assert_equals: expected "http://web-platform.test:8001/common/blank.html" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/url.window.html"
FAIL document.open() does not change document's URL (non-active document with an associated Window object; frame is removed) assert_equals: expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/url.window.html"
FAIL document.open() does not change document's URL (non-active document with an associated Window object; navigated away) assert_equals: expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/url.window.html"
FAIL document.open() does not change document's URL (non-active document without an associated Window object) assert_equals: expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/url.window.html"
Harness: the test ran to completion.
